study_materials_backend/study_materials/serializers.py
from rest_framework import serializers
from .models import Topic, StudyMaterial

class TopicSerializer(serializers.ModelSerializer):
    class Meta:
        model = Topic
        fields = '__all__'

class StudyMaterialSerializer(serializers.ModelSerializer):
    class Meta:
        model = StudyMaterial
        fields = '__all__'
